Use local ./node_path or node_modules/node
Common pattern is to use node as a devDependency in an npm package so that everything uses that version
of node; if it's present, run-node should probably use that as well.

Unless told not to. Which can be done by specifying the RUN_NODE_CACHE_PATH, in which case that file
will be sourced to update the PATH.

Or, you can do so by having a local .node_path file, which will be sourced if present (and RUN_NODE_CACHE_PATH
is not defined).